+++ NEWS.md
@@ -1,7 +1,7 @@
 # News about the `TcGSA` R package
 
 ### Main changes in Version 0.12.3 (2019-04-25) --- *this is only a minor release*:
-* prevent the simultaneous use of `baseline` and `group.var` in ploting functions 
+* prevent the simultaneous use of `baseline` and `group.var` in plotting functions 
 
 ### Main changes in Version 0.11.0 (2018-06-07) --- *this is only a minor release*:
 * Added a vignette